% Generated by build script
\name{omb_gdp_pct}
\alias{omb_gdp_pct}
\title{Fetch budget as percentage of GDP (Table 1.2)}
\usage{
omb_gdp_pct()
}
\arguments{
\item{url}{Optional URL to the HIST.xlsx file}
}
\value{tibble: fiscal_year, receipts_pct_gdp, outlays_pct_gdp,}
\description{Fetch budget as percentage of GDP (Table 1.2)}
