% Generated by build script
\name{omb_outlays_by_agency}
\alias{omb_outlays_by_agency}
\title{Fetch outlays by agency (Table 4.1)}
\usage{
omb_outlays_by_agency()
}
\arguments{
\item{url}{Optional URL to the HIST.xlsx file}
}
\value{tibble: category (agency name), fiscal_year, value, table_id}
\description{Fetch outlays by agency (Table 4.1)}
