% Generated by build script
\name{omb_summary}
\alias{omb_summary}
\title{Fetch budget summary (Table 1.1)}
\usage{
omb_summary()
}
\arguments{
\item{url}{Optional URL to the HIST.xlsx file}
}
\value{tibble: fiscal_year, receipts, outlays, surplus_deficit, table_id}
\description{Fetch budget summary (Table 1.1)}
